requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==2.5.1
sqlalchemy>=1.4,<2.0
psycopg2-binary
gunicorn==21.2.0
bleach==6.1.0
PyJWT==2.8.0
